+++ swing/common-ui/src/test/java/is/codion/swing/common/ui/component/table/FilterTableTest.java
@@ -20,7 +20,7 @@ package is.codion.swing.common.ui.component.table;
 
 import is.codion.swing.common.model.component.table.FilterTableModel;
 import is.codion.swing.common.model.component.table.FilterTableModel.TableSelection;
-import is.codion.swing.common.model.component.table.FilterTableSortModel;
+import is.codion.swing.common.model.component.table.FilterTableSorter;
 import is.codion.swing.common.ui.component.table.DefaultFilterTableSearchModel.DefaultRowColumn;
 import is.codion.swing.common.ui.component.table.FilterTableSearchModel.RowColumn;
 
@@ -257,7 +257,7 @@ public class FilterTableTest {
 		rowColumn = searchModel.results().next().orElse(null);
 		assertNull(rowColumn);
 
-		table.model().sort().descending(1);
+		table.model().sorter().descending(1);
 
 		searchModel.searchString().set("b");
 		rowColumn = searchModel.results().next().orElse(null);
@@ -284,7 +284,7 @@ public class FilterTableTest {
 						new DefaultRowColumn(3, 1)
 		), searchModel.results().get());
 
-		table.model().sort().ascending(1);
+		table.model().sorter().ascending(1);
 		table.columnModel().moveColumn(1, 0);
 
 		testModel.items().refresh();
@@ -301,7 +301,7 @@ public class FilterTableTest {
 		rowColumn = searchModel.results().next().orElse(null);
 		assertNull(rowColumn);
 
-		table.model().sort().descending(0);
+		table.model().sorter().descending(0);
 
 		searchModel.searchString().set("b");
 		rowColumn = searchModel.results().next().orElse(null);
@@ -346,10 +346,10 @@ public class FilterTableTest {
 		FilterTableModel<TestRow, Integer> tableModel = table.model();
 		AtomicInteger actionsPerformed = new AtomicInteger();
 		Runnable consumer = actionsPerformed::incrementAndGet;
-		table.model().sort().observer().addListener(consumer);
+		table.model().sorter().observer().addListener(consumer);
 
 		tableModel.items().refresh();
-		FilterTableSortModel<TestRow, Integer> sortModel = table.model().sort();
+		FilterTableSorter<TestRow, Integer> sortModel = table.model().sorter();
 		sortModel.order(0).set(SortOrder.DESCENDING);
 		assertEquals(SortOrder.DESCENDING, sortModel.columns().get(0).sortOrder());
 		assertEquals(E, tableModel.items().visible().get(0));
@@ -382,7 +382,7 @@ public class FilterTableTest {
 		sortModel.order(0).set(SortOrder.DESCENDING);
 		assertEquals(tableModel.items().visible().count() - 2, tableModel.items().visible().indexOf(NULL));
 		sortModel.order(0).set(SortOrder.UNSORTED);
-		table.model().sort().observer().removeListener(consumer);
+		table.model().sorter().observer().removeListener(consumer);
 	}
 
 	@Test
@@ -390,7 +390,7 @@ public class FilterTableTest {
 		FilterTable<TestRow, Integer> table = createTestTable(Comparator.reverseOrder());
 		FilterTableModel<TestRow, Integer> tableModel = table.model();
 		tableModel.items().refresh();
-		FilterTableSortModel<TestRow, Integer> sortModel = table.model().sort();
+		FilterTableSorter<TestRow, Integer> sortModel = table.model().sorter();
 		sortModel.order(0).set(SortOrder.ASCENDING);
 		assertEquals(E, tableModel.items().visible().get(0));
 		sortModel.order(0).set(SortOrder.DESCENDING);
@@ -416,13 +416,13 @@ public class FilterTableTest {
 		assertEquals(3, selectionModel.getMinSelectionIndex());
 		assertEquals(ITEMS.get(2), selectionModel.item().get());
 
-		table.model().sort().ascending(0);
+		table.model().sorter().ascending(0);
 		assertEquals(ITEMS.get(2), selectionModel.item().get());
 		assertEquals(2, selectionModel.getMinSelectionIndex());
 
 		tableModel.selection().indexes().set(singletonList(0));
 		assertEquals(ITEMS.get(0), selectionModel.item().get());
-		table.model().sort().descending(0);
+		table.model().sorter().descending(0);
 		assertEquals(4, selectionModel.getMinSelectionIndex());
 
 		assertEquals(singletonList(4), selectionModel.indexes().get());
@@ -604,7 +604,7 @@ public class FilterTableTest {
 						.collect(Collectors.toList());
 
 		model.items().add(rows);
-		model.sort().ascending(0);
+		model.sorter().ascending(0);
 		model.items().add(new TestRow("200"));
 
 		JViewport viewport = parentOfType(JViewport.class, table);
@@ -621,7 +621,7 @@ public class FilterTableTest {
 		testRow = model.items().visible().get(row);
 		assertEquals("301", testRow.value);
 
-		model.sort().clear();
+		model.sorter().clear();
 
 		model.items().visible().add(0, new TestRow("400"));
 		row = table.rowAtPoint(viewport.getViewPosition());